EC2 task networking for Windows: Build vpc-eni CNI plugin for Windows (#2893)
* Added CNI plugin version upgrade code to common file

* Build vpc-eni cni plugin for Windows

* Updated the git submodule for vpc-cni plugins

* Disabled verbose option for Windows unit tests

* Updated git submodule vpc-cni-plugins to latest release

* Updated values for max retry during network setup
