+++ .github/workflows/windows.yml
@@ -20,4 +20,4 @@ jobs:
         $Env:GOPATH = "$Env:GITHUB_WORKSPACE"
         cd "$Env:GITHUB_WORKSPACE"
         cd "src/github.com/aws/amazon-ecs-agent"
-        go test -v -race -tags unit -timeout 40s ./agent/...
+        go test -race -tags unit -timeout 40s ./agent/...
